PR-6a: the ensemble turned on — and the closure tiers say it is NOT ready
Ships: the member-ESS diagnostic (PLAN §6.4, behind a static
lss_member_diagnostics flag using the existing *_diagnostics wrapper
mechanism; goldens bit-identical with it off), the P13 consumed-budget
identity pinned against the integral the likelihood ACTUALLY evaluates,
the full §4.4 provenance rewiring, and the permitted/refused guard matrix
(35 rows, 41 tests, 11/11 independently re-verified against real argv).

P13 closes to 1e-15, and the survey-global normalizer is member-independent
BIT-EXACTLY — PLAN §4.2's gauge fixing, confirmed at the formation site.

**THE CLOSURE TIERS DO NOT PASS, AND PR-6a MUST NOT SHIP ON THEM:**

  Tier A  field recovery          PASS   slope 0.99908 +- 0.00433, r 0.99945
  Tier B  single-realization H0   FAIL   H0_true outside the 90% CI in 4/4 arms
  Tier C  coverage (n=24)         FAIL   KS p=0.0111, 9/24 outside 99%
  Tier D  misspecification x4     PASS   worst +0.390 sigma against a 0.5 gate

The decomposition is what matters, and it exonerates the field:
Tier B's 1.571 sigma latent-vs-table gap is 91% the **f_p channel**
(1.674 sigma); the field-only shift is 0.136 sigma and the table-field
shift 0.118 sigma. Tier C's overconfidence is x2.59 for latent and x2.26
for latent_off — i.e. it is NOT latent-specific. A bisection over five
configurations finds every f_p-aware rung puts H0_true inside its 90% CI:
the failure is dispersion, not any one feature.

Three source defects found, none fixed here (they need their own rung):

* S-1 count_map_solve's undamped fixed-trip Fisher iteration DIVERGES on
  5 of 8 nside-16 realizations (grad_inf to 5.0e5, period-2 cycle).
  Canonical link => Fisher == observed Hessian => this is exact Newton and
  needs a line search. Loud, not silent: the builder gates at 1e-8 and the
  production anchor converged at 1.09e-10.
* S-2 PLAN §3.4's b_gal rank-1 draw-covariance inflation is NOT
  IMPLEMENTED — `s_b` appears nowhere in darksirens/, laplace_draws takes
  only H_chol, and sensitivity_S's b_gal column is built, stored and never
  consumed. This makes Tier B's "latent CI >= table CI" gate vacuous and is
  the leading suspect for Tier C's x2.59 overconfidence.
* S-3 under c_mode=selection, off-footprint pixels are modelled as
  C_bar-COMPLETE unless f_p supplies the zero. 16/16 footprint runs without
  --per_pixel_completeness rail H0 to 125-138 at CDF 0.0000. Since
  loaders.py:1021 refuses f_p alongside a Q table, EVERY Q-table
  configuration on a footprint-limited survey is exposed — including
  production selq_radial. Pre-existing; not introduced here.

Also fixed here: eq. (4)'s closure floor on the shipped anchor was the f_p
STORAGE DTYPE — the builder formed (A, B, F_F) from the float64 map while
catalog_views hands the integrators float32. 1.6e-9 to 3.25e-8 growing with
b_GW, against 8.9e-16 when they agree. Same class as PR-5's fix, one input
over, and small enough to sit inside factory's own 1e-6 f_p guard.

Measured overhead (H100 NVL, 12 timed evals/arm): the seam + 8-member
ensemble costs +1.39% against the f_p baseline. 97.2% of PR-6a's headline
+95.4% is PR-2's --per_pixel_completeness, not the field. Correction: PR-0's
"3027 ms H100" baseline is an A100-80 number; the H100 value is 1417 ms, so
PLAN §2's percentages deflate 52x, not 110x.

Determinism: 100 repeats bit-identical in both arms; the production sweep
passes for latent (2.539) and FAILS for table (21.65). Stated in the report
in the same breath: under CRN both halves pass by construction and pass
equally on a distorted surrogate. Regression guard, never validation.

NO PRODUCTION 259-EVENT H0 RUN WAS LAUNCHED. It remains held for the owner.

+++ darksirens/cli/build_latent_field.py
@@ -123,9 +123,21 @@ def main(argv=None):
             f"occupancy guard 7: shells "
             f"{np.where(~(gal_ok & pix_ok))[0].tolist()} under-occupied; "
             f"reduce --n-shells.")
+    # float32 ROUND-TRIP, deliberately.  ``likelihood/catalog_views.py`` stores
+    # the run's ``f_p_rows`` / ``field_f_p_occ`` as float32, and those are what
+    # both integrators consume; forming ``(A, B, F_F)`` from the float64 map
+    # would normalize eq. (4) against a completeness the likelihood never sees.
+    # Measured on the shipped anchor: the mismatch leaves the budget identity
+    # closing at 1.6e-9 (b_GW = 0.37) to 3.25e-8 (b_GW = 3.77), growing with
+    # b_GW as e^{b f} predicts, against 8.9e-16 when the two agree.  Exactly the
+    # defect PR-5 fixed on the OTHER eq. (2) input (moments from the f64 draws
+    # rather than the stored f32 row factors) -- same class, one input over, and
+    # small enough to sit inside factory's own 1e-6 f_p guard, so nothing
+    # refuses such a run.  PLAN 4.2 claims eq. (4) is exact; this is what makes
+    # that true rather than true-to-1e-8.
     f_p = np.maximum(
         load_selection_fraction(args.per_pixel_completeness, nside).f_p[occ],
-        1e-3)
+        1e-3).astype(np.float32).astype(np.float64)
     print(f"[data] {occ.size} pixels, {int(counts.sum())} galaxies, "
           f"{args.n_shells} shells", flush=True)
